#28 Separated the upgrade files
