+++ classes/migration/install/PLNPluginSchemaMigration.inc.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * @file classes/PLNPluginSchemaMigration.inc.php
+ * @file classes/migration/install/PLNPluginSchemaMigration.inc.php
  *
  * Copyright (c) 2014-2022 Simon Fraser University
  * Copyright (c) 2000-2022 John Willinsky
