+++ upgrade.xml
@@ -13,6 +13,7 @@
 
 <install version="2.0.4.3">
 	<upgrade>
-		<migration class="plugins.generic.pln.PLNPluginSchemaUpgrade" />
+		<migration class="plugins.generic.pln.classes.migration.upgrade.I35_FixMissingField" />
+		<migration class="plugins.generic.pln.classes.migration.upgrade.I28_FixDepositStatus" />
 	</upgrade>
 </install>
